requirements.txt
eventlet
flask
flask_socketio
gevent
gevent-websocket
json
pymongo
python-dotenv
requests